@@ -1,6 +1,10 @@
 //------------------------------------------------------------------------------
 // <auto-generated>
-//     Generated by the MSBuild WriteCodeFragment class.
+//     This code was generated by a tool.
+//     Runtime Version:4.0.30319.42000
+//
+//     Changes to this file may cause incorrect behavior and will be lost if
+//     the code is regenerated.
 // </auto-generated>
 //------------------------------------------------------------------------------
 
@@ -14,3 +18,6 @@ using System.Reflection;
 [assembly: System.Reflection.AssemblyProductAttribute("invoicing.server")]
 [assembly: System.Reflection.AssemblyTitleAttribute("invoicing.server")]
 [assembly: System.Reflection.AssemblyVersionAttribute("1.0.0.0")]
+
+// Generated by the MSBuild WriteCodeFragment class.
+
